reafactor(id-generator): Moving ID generator util (#549)
* refactor(id-generator) File path changed

* refactor(id-generator) File path changed
