+++ packages/geo/src/lib/datasource/utils/index.ts
@@ -1,3 +1,2 @@
 export * from './esri-style-generator';
-export * from './id-generator';
 export * from './tilegrid';
