+++ packages/geo/src/lib/utils/id-generator.ts
@@ -2,10 +2,10 @@ import { Md5 } from 'ts-md5';
 
 import { uuid } from '@igo2/utils';
 
-import { AnyDataSourceOptions } from '../shared/datasources/any-datasource.interface';
-import { DataSourceOptions } from '../shared/datasources/datasource.interface';
-import { WMSDataSourceOptions } from '../shared/datasources/wms-datasource.interface';
-import { WMTSDataSourceOptions } from '../shared/datasources/wmts-datasource.interface';
+import { AnyDataSourceOptions } from '../datasource/shared/datasources/any-datasource.interface';
+import { DataSourceOptions } from '../datasource/shared/datasources/datasource.interface';
+import { WMSDataSourceOptions } from '../datasource/shared/datasources/wms-datasource.interface';
+import { WMTSDataSourceOptions } from '../datasource/shared/datasources/wmts-datasource.interface';
 
 /**
  * Generate a id from it's datasource options.
