+++ packages/geo/src/lib/utils/index.ts
@@ -1,2 +1,3 @@
 export * from './googleLinks';
+export * from './id-generator';
 export * from './osmLinks';
